fix: refactor delete message

+++ runtime/plaid/src/data/sqs/mod.rs
@@ -61,9 +61,7 @@ impl SQS {
             logger,
         }
     }
-}
 
-impl SQS {
     pub async fn drain_queue(&mut self) -> Result<(), String> {
         debug!("sqs/{} draining queue", self.config.name);
 
@@ -100,14 +98,7 @@ impl SQS {
                         if let Some(id) = message.message_id() {
                             if self.seen_messages.contains(id) {
                                 debug!("sqs/{} detected duplicate message {id}", self.config.name);
-                                if let Some(receipt_handle) = message.receipt_handle {
-                                    if let Ok(_) = self.delete_message(receipt_handle).await {
-                                        debug!(
-                                            "sqs/{} deleted_message {:?}",
-                                            self.config.name, message.message_id,
-                                        );
-                                    }
-                                }
+                                self.delete_message(message.receipt_handle).await?;
                                 continue;
                             } else {
                                 self.seen_messages.put(id.to_string(), 0u32);
@@ -118,14 +109,7 @@ impl SQS {
                             // send event to rules
                             self.send_for_processing(body.as_bytes().to_vec())?;
                             // delete the message from the queue to prevent re-processing
-                            if let Some(receipt_handle) = message.receipt_handle {
-                                if let Ok(_) = self.delete_message(receipt_handle).await {
-                                    debug!(
-                                        "sqs/{} deleted_message {:?}",
-                                        self.config.name, message.message_id,
-                                    );
-                                }
-                            }
+                            self.delete_message(message.receipt_handle).await?;
                         }
                     }
                 }
@@ -133,15 +117,19 @@ impl SQS {
         }
     }
 
-    async fn delete_message(&self, receipt_handle: String) -> Result<(), String> {
-        let _ = self
-            .client
-            .delete_message()
-            .queue_url(&self.config.queue_url)
-            .receipt_handle(receipt_handle)
-            .send()
-            .await
-            .map_err(|e| format!("sqs/{} delete_message failed: [{e}]", self.config.name))?;
+    async fn delete_message(&self, receipt_handle: Option<String>) -> Result<(), String> {
+        if let Some(receipt_handle) = receipt_handle {
+            let _ = self
+                .client
+                .delete_message()
+                .queue_url(&self.config.queue_url)
+                .receipt_handle(receipt_handle)
+                .send()
+                .await
+                .map_err(|e| format!("sqs/{} delete_message failed: [{e}]", self.config.name))?;
+
+            debug!("sqs/{} deleted_message", self.config.name,);
+        }
         Ok(())
     }
